Contracts_App_Pro/requirements.txt
PyQt6
pandas
openpyxl
requests
python-docx
reportlab
pywin32
